ordb-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
pydantic==2.5.2
